Add base infrastructure for site.
Terraform configuration for the primary infrastructure required for the site.
The site will be primarily a static site served out of a storage bucket.
There will also be an API that is served by App Engine.

Network configuration will redirect all http traffic to https.
Direct requests to the App Engine appspot URLs are blocked, so all
traffic to the API must pass through the loadbalancer.

Some entities that can't be applied until App Engine is initialized
and services deployed to it have been commented out, but left in
for illustrative purposes.