Add initial Cloud Armor security policy.
I don't want people, scanners, bots, etc. reaching the site by its IP.
This commit adds a Cloud Armor security policy that denies all traffic by default
and only accepts traffic that has a host header that contains the domain.

This also changes up the backends so that the default service directs to the Cloud Run
backend, so that I can apply the header inspection rules to that traffic.

This isn't foolproof since headers can be forged, but it'll catch things that are just out
there scanning IPs. The next step will be to create a blackhole service to catch and log
any traffic that doesn't match an explicitly handled host at the load balancer.