Set up blackhole storage bucket.
The cloud logs show some scanners accessing the site by its IP.
I would prefer to just drop all traffic that doesn't come in for a
hostname that I explicitly serve, but Google Cloud load balancing with
Serverless NEGs doesn't appear to allow me to do that. I have to specify
a default service for unmatched traffic, and that at the very least responds
with a 404.

I'm going to keep researching to see if there's a way to accomplish this,
but in the meantime, I am setting up a second bucket for static content.
This bucket only contains a 404 page, and is the default backend for the load
balancer.

This allows me to direct all requests that don't match a host that I
am intentionally serving away from the backends that serve real content.
Instead, all this traffic will be routed to this blackhole bucket, which
will always just return a very basic 404 page.

People can always still scan the site by crawling around the actual hosts,
but this at least directs maliciious traffic with forged or missing host headers
away from my real services.