Initial next.js implementation.
I decided to switch the frontend to next.js. I like the ease of publishing with a pure static site generator, but I'd like to have some more dynamic elements and an API, and it was pretty clunky to try to shim that in to a template-driven SSG. Next.js gives me more flexibility to build the sort of site I'm imagining at the cost of having to reimplement the basic markdown-to-html SSG capability. But it's a fun learning project and it's nice not to be 20 years behind the times on front-end web technologies any more.

Now I need to tweak the GCloud stuff to send everything to Cloud Run and ditch the static bucket.