Deploy next.js app to cloud run.
Dockerize the next.js app so that it can be deployed to cloud run.
Reconfigure the load balancer to direct traffic for the bare domain to the new cloud run service (web).

I'm no longer using the static content bucket. I'll remove that in a future commit.
I will leave the bucket for the blackhole in place so that requests that don't match any
served URLs are routed away from cloud run.

This commit just has the quick changes needed to dockerize the service and set up
the google cloud infrastructure correctly. There'll be some cleanup commits to follow.