User-managed service account for cloud run deployment.
Create a user-managed service account for the 'web' cloud run service.
This account has no permissions, so it can't do anything in google cloud.
It will never need to, since all this service does is serve the next.js app.

Any future functionality that interacts with GCP entities will be in other services
that compose the API.

Got rid of a no-longer-used variable while I was at it. I don't need to know the services in terraform just now.

I may in the future if I want to create the service accounts from variables.