Configure Cloud CDN to use Cache-Control headers to determine caching behavior.
This was fun. When I was trying out cloud CDN, I set it up to Cache all static content,
which is the default. That worked, but I'd prefer to have control over the caching behavior.
Next.js allows you to define the cache-control headers, so that's the way to go. But I couldn't
change the Cloud CDN caching type in terraform. Whenever I tried to change it fro CACHE_ALL_STATIC to
USE_ORIGIN_HEADERS, terraform would fail because it kept sending along the old ttl settings
(which aren't valid when the cache control headers are setting the TTLs).

In the end, I had to go scorched-earth and delete the backend entirely to get this to work.
I decided to take the opportunity to rename the backend while I was at it.