Use the next.js router to determine the current section.
Instead of passing the section into the Layout component everywhere,
use what I learned yesterday about the router to determine which section
we're in based on the path.

This simplifies the components a good deal.

Next step: use the router to determine whether or not I'm on an index page for a section.
Then I won't have to pass anything into the Layout.