Reimplement SiteMenu with MUI components
I'm trying out moving over to MUI components for the majority of the UI.
I don't need to be implementing a full suite of custom components, and I thought this might save some time.

That said, so far it hasn't. I think the horizontal menu isn't really something MUI supports.
I can have an AppBar with buttons, and I can pretty easily implememt a vertical menu that shows the active
item. But doing this as a horizontal menu stil ended up being a bunch of messing with divs.

That said, hopefully I benefit some from the standard typography components and whatnot.
I also need ot see how much more of this I can implement as theme overrides so all the color definitions
are in one place.

But for now, this commit has the menu bar looking and acting as it did before using all 3rd-party components.