Simplify Site Menu
Convert the menu to entirely MUI components.
I'm able to do all the dimension-fudging that I was trying to do before to make sure that links
used the full width of the menu bar just by using the default tab layout.

Using tabs with the component set to the Next.js Link component preserves the client-side routing
behavior as well.

This is much simpler and cleaner, and is all I need for what I'm trying to accomplish.

The dynamic menu generation may be a bit of overkill, but I do have a few more sections in mind,
so it'll be nice to just pick them up in the menu as I add their directories.